+++ src/components/BurgerIngredients/index.ts
@@ -1,0 +1,1 @@
+export { BurgerIngredients } from './BurgerIngredients';
