+++ src/components/app/App.jsx
@@ -1,16 +1,16 @@
 import React from 'react';
-import appStyles from './app.module.css';
-import { AppHeader } from '../app-header'
-import { BurgerConstructor } from '../BurgerConstructor'
-import { BurgerIngredient } from '../BurgerIngredients'
+import appStyles from './App.module.css';
+import { AppHeader } from '../AppHeader'
+import { BurgerIngredients } from '../BurgerIngredients'
+import { BurgerConstructor } from '../BurgerConstructors'
 
 export function App() {
   return (
     <div className={appStyles.app}>
       <AppHeader />
       <main className={appStyles.container}>
+        <BurgerIngredients />
         <BurgerConstructor />
-        <BurgerIngredient />
       </main>
     </div>
   );
